Move various files out of blockfolder and into their own files. Start building the LSTM block for testing
