+++ training/configs/model/modules/blocks/conv_lstm_block.yaml
@@ -1,0 +1,5 @@
+_target_: training.dlwp.model.modules.blocks.ConvLSTMBlock
+_recursive_: false
+in_channels: 3
+kernel_size: 1
+downscale_factor: 4
