+++ training/configs/model/modules/blocks/conv_next_block.yaml
@@ -6,5 +6,6 @@ _recursive_: true
 in_channels: 3
 out_channels: 1
 kernel_size: 3
+latent_conv_size: 1
 dilation: 1
 upscale_factor: 4
