+++ training/configs/model/modules/blocks/learned_pool.yaml
@@ -1,0 +1,2 @@
+_target_: training.dlwp.model.modules.blocks.pool.LearnedPool
+pooling: 2
