+++ training/configs/model/modules/blocks/transposed_conv_upsample.yaml
@@ -1,7 +1,7 @@
 defaults:
   - /model/modules/activations@activation: capped_gelu
 
-_target_: training.dlwp.model.modules.blocks.TransposedConvUpsample
+_target_: training.dlwp.model.modules.blocks.interpolate.TransposedConvUpsample
 in_channels: 3
 out_channels: 1
 upsampling: 2
